requirements-production.txt
django
django-bootstrap3
django-registration
django-phonenumber-field
django-related-admin
django-widget-tweaks
celery
django-celery
python-dateutil
pytz

# for graph_models
# TODO: put this in -dev when we figure out how to split settings file
django-extensions
